requirements.txt
attrdict==2.0.1
imageio==2.6.1
matplotlib==3.1.1
numpy==1.16.4
Pillow==6.2.1
scikit-image==0.16.1
scipy==1.4.1
tqdm==4.36.1
yacs==0.1.6
tensorboard==2.1.1
